Convert indent double-space to tabs
The google apps script editor uses double-space and having tabs in this
code is then annoying, however, the Zotero codebase uses tabs across all
repos and it is difficult to get editor settings that will treat this one
file differently. There shouldn't be as much editing in the apps script
editor going forward, so we can go with tabs.