A second attempt at preventing concurrent citation editing
The attempt in (1dbbf6cf) does not work, because we cannot acquire
document locks via the Apps Script LockService since we publish the
code as API executable, not a proper add-on.

This instead stores a "Z_LOCK" NamedRange in the document while it is
being edited that other invocations of the script must respect. If the
lock fails to clear for some reason, the locked document prompt allows
a manual lock override.