Do not wait for gdocs to save changes when inserting a field
This was not strictly necessary until the user actually selects
some references and presses enter in the citation dialog, and by then
it should have saved the changes. If the changes are not saved
insertion might fail.

The change is due to numerous reports of inserts recently being very
slow on Chrome. e.g. https://forums.zotero.org/discussion/74464/zotero-google-docs-is-very-very-slow