Improve selected field detection
Instead of adding a space and removing it to see whether the link
popover appears we open insert link popover, check whether there is a
link in the URL field and close it by clicking Apply (or Escape, if no
link present). This is done because in some recent Google Docs change
the link popover no longer appears when the cursor is at the right edge
of the citation.

Report: https://forums.zotero.org/discussion/87006/google-docs-bug