Fix insertions on Mac Edge resulting in plaintext citations
Regression from 2ffbb4f6.

We're trying to find a way to trigger the google docs sync icon after
running an apps script command in the background so we're sure we have
up-to-date text before attempting to move the cursor to the end of the
citation. Adding and removing a character at the wrong time (especially
if you add a spacebar before receiving apps-script updates and remove it
after) seems to remove the link in text.

This instead triggers the bold text modifier on and off, which triggers
gdocs sync without messing up the link in text.