Implement saving for non-translated website resources
This patch has 2 components. First, when we initially load the page, we detect whether we have a translator for the page or not. Before this patch, untranslated pages had their import button disabled. After the patch, we annotate the tab to indicate we don't have a translator, and leave the button visible.

Second, when the user interacts with the import button, we check to see if no translators were found, and if so we import the website as a simple URL reference, following the bibtex/biblatex standard. This is modelled after the savePdf functionality.

I have tested this locally with translated and non-translated websites. Note that I did change the savePdf function to use the urlDate field instead of the accessDate field, in order to match the bibtex standard. JabRef now correctly shows the URL access date on imported PDFs and simple webpages.